fix(ColorTextBox): use variables; sort directly under atoms; vertically align
